Logic Moved to another project
