+++ Game/Logic/IGameControl.cs
@@ -1,12 +1,4 @@
 ﻿namespace Game.Logic
 {
-    internal interface IGameControl
-    {
-        public void PlayerMoveLeft();
-        public void PlayerMoveRight();
-        public void PlayerJump();
-        public void PlayerCrouch();
-        public void PlayerMelee();
-        public void PlayerAttack();
-    }
+    
 }
