+++ Game/Logic/IGameModel.cs
@@ -3,9 +3,5 @@ using System.Windows.Media;
 
 namespace Game.Logic
 {
-    internal interface IGameModel
-    {
-        public string GetLevelPath();
-        public Player GetPlayer();
-    }
+
 }
