+++ Grand Train Robbery.sln
@@ -11,6 +11,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GrandTrainRobbery.Models", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GrandTrainRobbery.Physics", "GrandTrainRobbery.Physics\GrandTrainRobbery.Physics.csproj", "{80EED018-E0A8-41E5-AFD1-01DC0F3ED00A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GrandTrainRobbery.Logic", "GrandTrainRobbery.Logic\GrandTrainRobbery.Logic.csproj", "{87440E96-3D67-49E2-8E62-6DB3999F9854}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{80EED018-E0A8-41E5-AFD1-01DC0F3ED00A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{80EED018-E0A8-41E5-AFD1-01DC0F3ED00A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{80EED018-E0A8-41E5-AFD1-01DC0F3ED00A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{87440E96-3D67-49E2-8E62-6DB3999F9854}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{87440E96-3D67-49E2-8E62-6DB3999F9854}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{87440E96-3D67-49E2-8E62-6DB3999F9854}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{87440E96-3D67-49E2-8E62-6DB3999F9854}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
