+++ GrandTrainRobbery.Logic/GrandTrainRobbery.Logic.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\GrandTrainRobbery.Models\GrandTrainRobbery.Models.csproj" />
+    <ProjectReference Include="..\GrandTrainRobbery.Physics\GrandTrainRobbery.Physics.csproj" />
+  </ItemGroup>
+
+</Project>
